backbone_example/requirements.txt
Django==1.3.2
mimeparse>=0.1.3
python-dateutil==1
django-tastypie==0.9.11
pystache==0.3.1
django-guardian==1.0.4
django-extensions==0.9
ipdb==0.7
ipython==0.13
lazy==1.0